Fix/marker zoom (#7)
* Fix marker zoom issue

+++ src/core.ts
@@ -116,7 +116,7 @@ function buildMarkerData(file: VFile): void {
       position: { x: parseInt(marker.x), y: parseInt(marker.y) },
       icon: marker.icon,
       colour: getColourValue(marker.colour),
-      minZoom: marker.minZoom ? parseInt(marker.minZoom) : undefined,
+      minZoom: marker.minZoom ? parseFloat(marker.minZoom) : undefined,
     });
   }
 }
